requirements.txt
streamlit==1.32.0
openai==1.12.0
python-dotenv==1.0.1
python-docx==1.0.1
PyPDF2==3.0.1
pdfplumber==0.10.3 